update writer with data class property.
Separating writing of groups, data and metadata to avoid confusion

I have the feeling the writers can benefit for another cleanup round.
But pushing this first PR for you to accept the mechanics

+++ geoh5py/data/float_data.py
@@ -46,7 +46,7 @@ class FloatData(NumericData):
         if values is None:
             return values
 
-        return values.astype(np.float32)
+        return values
 
     @classmethod
     def primitive_type(cls) -> PrimitiveTypeEnum:
